+++ step_3_repeat.go
